split up and unclutter LoRAGPT2Implementation.py
